fed_yolo/requirements.txt
absl-py==0.11.0
amqp==2.6.1
billiard==3.6.3.0
cachetools==4.2.1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
dill==0.3.3
distro-info===0.23ubuntu1
docker-pycreds==0.4.0
et-xmlfile==1.0.1
Flask==1.1.2
gevent==20.6.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.26.1
google-auth-oauthlib==0.4.2
greenlet==1.0.0
grpcio==1.35.0
gunicorn==20.0.4
h5py==3.1.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
joblib==1.0.0
kiwisolver==1.3.1
kombu==4.6.11
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mpi4py==3.0.3
networkx==2.5
numexpr==2.7.2
numpy==1.20.0
oauthlib==3.1.0
opencv-python==4.5.1.48
paho-mqtt==1.5.0
pathtools==0.1.2
Pillow==8.1.0
promise==2.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.2.0
PyGObject==3.36.0
pyparsing==2.4.7
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
sentry-sdk==0.19.5
setproctitle==1.2.2
Shapely==1.7.1
shortuuid==1.0.1
six==1.14.0
smmap==3.0.5
subprocess32==3.5.4
tables==3.6.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminal==0.4.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2021.2.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.56.0
typing-extensions==3.7.4.3
unattended-upgrades==0.1
urllib3==1.25.9
vine==1.3.0
wandb==0.10.17
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
